Extract Monty execution into dedicated tool module
